requirements.txt
openai>=1.13.3
python-dotenv==1.0.1
python-telegram-bot==21.0.1
requests~=2.31.0